Parsing file, creating rootdoc, passing to Doclet.
Basic functionality of the Comtor API is here!

Calling:
java COMTOR Doclet Source

where "Doclet" is a subclass of the abstract Doclet type that implements
the boolean "start" function, and "Source" is a compilably Java source
file or directory.

Note that the accompanying project, ComtorEngine, is needed to actually
use the COMTOR parser.